# Conflicts:
#	Assets/Scenes/Suzumura/Resources/CSV/enemy.csv.meta
#	Assets/Scenes/Suzumura/Resources/CSV/enemy01.csv.meta
#	Assets/Scenes/Suzumura/Stage Prehab/EnemyGroup.prefab.meta
#	Assets/Suzumura/Resources/CSV/enemy.csv
#	Assets/Suzumura/Resources/CSV/enemy.csv.meta
#	Assets/Suzumura/Resources/CSV/enemy00.csv
#	Assets/Suzumura/Resources/CSV/enemy00.csv.meta
#	Assets/Suzumura/Resources/CSV/enemy01.csv
#	Assets/Suzumura/Resources/CSV/enemy01.csv.meta
#	Assets/Suzumura/Resources/CSV/enemy11.csv
#	Assets/Suzumura/Resources/CSV/enemy11.csv.meta
#	Assets/Suzumura/Resources/CSV/enemy12.csv
#	Assets/Suzumura/Resources/CSV/enemy12.csv.meta
#	Assets/Suzumura/Resources/CSV/enemy2.csv
#	Assets/Suzumura/Resources/CSV/enemy2.csv.meta
#	Assets/Suzumura/Stage Prehab/EnemyGroup.prefab
#	Assets/Suzumura/Stage Prehab/EnemyGroup.prefab.meta
#	Assets/Suzumura/master Suzu.unity
#	Assets/Suzumura/master Suzu.unity.meta

+++ Assets/Scripts/Suzumura/EnemyF/EnemyData.cs
@@ -7,13 +7,13 @@ public class EnemyData
     public string name;                 // 名前(No.)
     public EnemyCategory category;      // カテゴリー
     public int HP;                      // HP
-    public int Size;                    // 大きさ
+    public float Size;                  // 大きさ
     public int StartPosX;               // 出現位置
     public int StartPosY;               // 
     public int StartPosZ;               // 
-    public int EntryPosX;               // 画面内に入る位置
-    public int EntryPosY;               //
-    public int EntryPosZ;               //
+    public float EntryPosX;             // 画面内に入る位置
+    public float EntryPosY;             //
+    public float EntryPosZ;             //
     public int sideNo;                  // 辺
     public int Target1PosX;             // 目標地点1
     public int Target1PosY;             //
